reindex_everything should ignore non-RDF sources Because they don't have parsable graphs.
