+++ spec/unit/base_spec.rb
@@ -20,37 +20,6 @@ describe ActiveFedora::Base do
     end
   end
 
-  describe "get_descendent_uris" do
-
-    before :each do
-      ids.each do |id|
-        ActiveFedora::Base.create id: id
-      end
-    end
-
-    def root_uri(ids=[])
-      ActiveFedora::Base.id_to_uri(ids.first)
-    end
-
-    context 'when there there are no descendents' do
-
-      let(:ids) { ['foo'] }
-
-      it 'returns an array containing only the URI passed to it' do
-        expect(ActiveFedora::Base.get_descendent_uris(root_uri(ids))).to eq ids.map {|id| ActiveFedora::Base.id_to_uri(id) }
-      end
-    end
-
-    context 'when there are > 1 descendents' do
-
-      let(:ids) { ['foo', 'foo/bar', 'foo/bar/chu'] }
-
-      it 'returns an array containing the URI passed to it, as well as all descendent URIs' do
-        expect(ActiveFedora::Base.get_descendent_uris(root_uri(ids))).to eq ids.map {|id| ActiveFedora::Base.id_to_uri(id) }
-      end
-    end
-  end
-
   describe "With a test class" do
     before :each do
       class FooHistory < ActiveFedora::Base
